:recycle: refactor: nest setup 분리
- 추후 e2e 테스트 코드 작성시 nest setup을 위해 분리
